+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.27
 - bug fixes in auto require and post classpath parameter bug when running in debug mode
+- wls_multi_datasource resource type added
 
 ## 1.0.26
 - auto require based on the wls resource parameters, no need to use require on the all wls resource types